+++ frontend_tests/node_tests/message_edit.js
@@ -9,6 +9,7 @@ const {page_params} = require("../zjsunit/zpage_params");
 page_params.realm_community_topic_editing_limit_seconds = 259200;
 
 const message_edit = zrequire("message_edit");
+const settings_config = zrequire("settings_config");
 
 const get_editability = message_edit.get_editability;
 const editability_types = message_edit.editability_types;
@@ -81,7 +82,8 @@ run_test("get_editability", () => {
         sent_by_me: false,
         type: "stream",
     };
-    page_params.realm_allow_community_topic_editing = true;
+    page_params.realm_edit_topic_policy =
+        settings_config.common_message_policy_values.by_everyone.code;
     page_params.realm_allow_message_editing = true;
     page_params.realm_message_content_edit_limit_seconds = 0;
     page_params.realm_community_topic_editing_limit_seconds = 259200;
@@ -93,15 +95,18 @@ run_test("get_editability", () => {
     assert.equal(message_edit.is_topic_editable(message), true);
 
     message.sent_by_me = true;
-    page_params.realm_allow_community_topic_editing = false;
+    page_params.realm_edit_topic_policy =
+        settings_config.common_message_policy_values.by_admins_only.code;
     assert.equal(message_edit.is_topic_editable(message), true);
 
     message.sent_by_me = false;
-    page_params.realm_allow_community_topic_editing = false;
+    page_params.realm_edit_topic_policy =
+        settings_config.common_message_policy_values.by_admins_only.code;
     assert.equal(message_edit.is_topic_editable(message), false);
 
     message.sent_by_me = false;
-    page_params.realm_allow_community_topic_editing = false;
+    page_params.realm_edit_topic_policy =
+        settings_config.common_message_policy_values.by_admins_only.code;
     page_params.is_admin = true;
     assert.equal(message_edit.is_topic_editable(message), true);