+++ static/js/admin.js
@@ -18,9 +18,6 @@ import * as settings_toggle from "./settings_toggle";
 
 const admin_settings_label = {
     // Organization settings
-    realm_allow_community_topic_editing: $t({
-        defaultMessage: "Users can edit the topic of any message",
-    }),
     realm_allow_edit_history: $t({defaultMessage: "Enable message edit history"}),
     realm_mandatory_topics: $t({defaultMessage: "Require topics in stream messages"}),
     realm_notifications_stream: $t({defaultMessage: "New stream notifications:"}),
@@ -82,7 +79,6 @@ export function build_page() {
         realm_avatar_changes_disabled: page_params.realm_avatar_changes_disabled,
         realm_add_emoji_by_admins_only: page_params.realm_add_emoji_by_admins_only,
         can_add_emojis: settings_emoji.can_add_emoji(),
-        realm_allow_community_topic_editing: page_params.realm_allow_community_topic_editing,
         realm_message_content_edit_limit_minutes: settings_org.get_realm_time_limits_in_minutes(
             "realm_message_content_edit_limit_seconds",
         ),