+++ static/js/message_edit.js
@@ -24,6 +24,7 @@ import * as overlays from "./overlays";
 import {page_params} from "./page_params";
 import * as resize from "./resize";
 import * as rows from "./rows";
+import * as settings_config from "./settings_config";
 import * as settings_data from "./settings_data";
 import * as stream_bar from "./stream_bar";
 import * as stream_data from "./stream_data";
@@ -68,7 +69,10 @@ export function is_topic_editable(message, edit_limit_seconds_buffer = 0) {
         return true;
     }
 
-    if (!page_params.realm_allow_community_topic_editing) {
+    if (
+        page_params.realm_edit_topic_policy ===
+        settings_config.common_message_policy_values.by_admins_only.code
+    ) {
         // If you're another non-admin user, you need community topic editing enabled.
         return false;
     }