+++ static/js/settings_config.js
@@ -198,6 +198,19 @@ export const wildcard_mention_policy_values = {
     },
 };
 
+export const common_message_policy_values = {
+    by_everyone: {
+        order: 1,
+        code: 5,
+        description: $t({defaultMessage: "Admins, members and guests"}),
+    },
+    by_admins_only: {
+        order: 2,
+        code: 2,
+        description: $t({defaultMessage: "Admins only"}),
+    },
+};
+
 const time_limit_dropdown_values = new Map([
     [
         "any_time",