+++ zerver/lib/events.py
@@ -218,8 +218,8 @@ def fetch_initial_state_data(
         state["realm_allow_message_editing"] = (
             False if user_profile is None else realm.allow_message_editing
         )
-        state["realm_allow_community_topic_editing"] = (
-            False if user_profile is None else realm.allow_community_topic_editing
+        state["realm_edit_topic_policy"] = (
+            Realm.POLICY_ADMINS_ONLY if user_profile is None else realm.edit_topic_policy
         )
         state["realm_allow_message_deleting"] = (
             False if user_profile is None else realm.allow_message_deleting